Fixed the threshold for OVEREXPOSED
git-svn-id: http://svnhq2.hq.eso.org/p2/trunk/Pipelines/cr2re/cr2rep@266435 318c8218-2dab-440d-9263-6f70bac63e5b

+++ cr2res/cr2res_qc.c
@@ -448,7 +448,7 @@ double cr2res_qc_overexposed(
     }
     cpl_image_delete(labels) ;
     if (nb_total == 0) return -1.0 ;
-    return nb_over / nb_total ;
+    return (double)nb_over/(double)nb_total ;
 }
 
 /*----------------------------------------------------------------------------*/
